requirements.txt
Faker==35.0.0
fastapi==0.115.7
langchain==0.3.15
langchain_chroma==0.2.1
langchain_community==0.3.15
langchain_core==0.3.31
langchain_openai==0.3.2
langchain_text_splitters==0.3.5
plotly==6.0.0
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.41.1
psycopg2-binary==2.9.10
streamlit==1.41.1
uvicorn==0.34.0
kaleido==0.4.2